Move TestUtils in separated file
